+++ trunk/src/main/java/org/ajmm/obsearch/index/pivotselection/RandomPivotSelector.java
@@ -49,7 +49,7 @@ public class RandomPivotSelector implements PivotSelector {
      * @see org.ajmm.obsearch.index.PivotSelector#generatePivots(short)
      */
     public void generatePivots(AbstractPivotIndex index ) throws OBException, IllegalAccessException, InstantiationException, DatabaseException{
-    	byte pivots = index.getPivotsCount();
+    	short pivots = index.getPivotsCount();
     	int maxIdAvailable = index.getMaxId();
         final int maxId = maxIdAvailable + 1;
         int[] res = new int[pivots];
